backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "black>=26.1.0",
    "coverage>=7.13.4",
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "isort>=8.0.0",
    "mypy>=1.19.1",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
